refactor: organize gestures
