+++ app/src/components/PianoRoll/MouseHandler/PencilMouseHandler.ts
@@ -1,18 +1,16 @@
-import {
-  eventsInSelection,
-  useSelectNote,
-  useStartNote,
-  useStopNote,
-} from "../../../actions"
 import { Point } from "../../../entities/geometry/Point"
-import { Selection } from "../../../entities/selection/Selection"
-import { NoteNumber } from "../../../entities/unit/NoteNumber"
-import { observeDrag2 } from "../../../helpers/observeDrag"
 import { useStores } from "../../../hooks/useStores"
 import { PianoNoteItem } from "../../../stores/PianoRollStore"
-import { isNoteEvent, NoteEvent } from "../../../track"
-import { useMoveDraggableGesture } from "./moveDraggableAction"
-import { MouseGesture } from "./NoteMouseHandler"
+import { useAddNoteToSelectionGesture } from "./gestures/useAddNoteToSelectionGesture"
+import { useCreateNoteGesture } from "./gestures/useCreateNoteGesture"
+import {
+  useDragNoteCenterGesture,
+  useDragNoteLeftGesture,
+  useDragNoteRightGesture,
+} from "./gestures/useDragNoteEdgeGesture"
+import { useRemoveNoteFromSelectionGesture } from "./gestures/useRemoveNoteFromSelectionGesture"
+import { useRemoveNoteGesture } from "./gestures/useRemoveNoteGesture"
+import { useSelectNoteGesture } from "./gestures/useSelectNoteGesture"
 
 export const usePencilGesture = () => {
   const { pianoRollStore } = useStores()
@@ -121,266 +119,4 @@ const getPositionType = (
     return "right"
   }
   return "center"
-}
-
-const useDragNoteEdgeGesture =
-  (edge: "left" | "right" | "center") => (): MouseGesture<[number]> => {
-    const {
-      pianoRollStore,
-      pianoRollStore: { selectedTrack, selectedNoteIds },
-    } = useStores()
-    const selectNote = useSelectNote()
-    const startNote = useStartNote()
-    const stopNote = useStopNote()
-    const moveDraggableAction = useMoveDraggableGesture()
-
-    return {
-      onMouseDown(e: MouseEvent, noteId: number) {
-        if (
-          selectedTrack === undefined ||
-          selectedTrack.channel === undefined
-        ) {
-          return
-        }
-
-        const note = selectedTrack.getEventById(noteId)
-        if (note == undefined || !isNoteEvent(note)) {
-          return
-        }
-
-        const isSelected = selectedNoteIds.includes(noteId)
-
-        if (!isSelected) {
-          selectNote(noteId)
-        }
-
-        const newSelectedNoteIds = pianoRollStore.selectedNoteIds
-
-        const { channel } = selectedTrack
-        startNote({ ...note, channel })
-        let playingNoteNumber = note.noteNumber
-
-        moveDraggableAction.onMouseDown(
-          e,
-          { type: "note", position: edge, noteId },
-          newSelectedNoteIds
-            .filter((id) => id !== noteId)
-            .map((noteId) => ({
-              type: "note",
-              position: edge,
-              noteId,
-            })),
-          {
-            onChange(_e, { oldPosition, newPosition }) {
-              const newNote = selectedTrack.getEventById(noteId)
-              if (newNote == undefined || !isNoteEvent(newNote)) {
-                return
-              }
-              // save last note duration
-              if (oldPosition.tick !== newPosition.tick) {
-                pianoRollStore.lastNoteDuration = newNote.duration
-              }
-              if (
-                oldPosition.noteNumber !== newPosition.noteNumber &&
-                newNote.noteNumber !== playingNoteNumber
-              ) {
-                stopNote({ noteNumber: playingNoteNumber, channel })
-                startNote({
-                  noteNumber: newNote.noteNumber,
-                  channel,
-                  velocity: newNote.velocity,
-                })
-                playingNoteNumber = newNote.noteNumber
-              }
-            },
-            onMouseUp() {
-              stopNote({ noteNumber: playingNoteNumber, channel })
-            },
-            onClick(e) {
-              if (!e.shiftKey) {
-                selectNote(noteId)
-              }
-            },
-          },
-        )
-      },
-    }
-  }
-
-const useDragNoteLeftGesture = useDragNoteEdgeGesture("left")
-const useDragNoteRightGesture = useDragNoteEdgeGesture("right")
-const useDragNoteCenterGesture = useDragNoteEdgeGesture("center")
-
-const useCreateNoteGesture = (): MouseGesture => {
-  const {
-    song: { timebase },
-    pianoRollStore,
-    pianoRollStore: { transform, selectedTrack, quantizer, newNoteVelocity },
-    pushHistory,
-  } = useStores()
-  const dragNoteCenterAction = useDragNoteCenterGesture()
-
-  return {
-    onMouseDown(e: MouseEvent) {
-      if (e.shiftKey) {
-        return
-      }
-
-      const local = pianoRollStore.getLocal(e)
-      const { tick, noteNumber } = transform.getNotePoint(local)
-
-      if (
-        selectedTrack === undefined ||
-        selectedTrack.channel == undefined ||
-        !NoteNumber.isValid(noteNumber)
-      ) {
-        return
-      }
-
-      pushHistory()
-
-      const quantizedTick = selectedTrack.isRhythmTrack
-        ? quantizer.round(tick)
-        : quantizer.floor(tick)
-
-      const duration = selectedTrack.isRhythmTrack
-        ? timebase / 8 // 32th note in the rhythm track
-        : (pianoRollStore.lastNoteDuration ?? quantizer.unit)
-
-      const note = selectedTrack.addEvent({
-        type: "channel",
-        subtype: "note",
-        noteNumber: noteNumber,
-        tick: quantizedTick,
-        velocity: newNoteVelocity,
-        duration,
-      } as NoteEvent)
-
-      if (note === undefined) {
-        return
-      }
-
-      dragNoteCenterAction.onMouseDown(e, note.id)
-    },
-  }
-}
-
-const useRemoveNoteGesture = (): MouseGesture => {
-  const {
-    pianoRollStore,
-    pianoRollStore: { selectedTrack },
-    pushHistory,
-  } = useStores()
-
-  const removeEvent = (eventId: number) => {
-    if (selectedTrack === undefined) {
-      return
-    }
-    pushHistory()
-    selectedTrack.removeEvent(eventId)
-    pianoRollStore.selectedNoteIds = pianoRollStore.selectedNoteIds.filter(
-      (id) => id !== eventId,
-    )
-  }
-
-  return {
-    onMouseDown(e: MouseEvent) {
-      const startPos = pianoRollStore.getLocal(e)
-      const items = pianoRollStore.getNotes(startPos)
-      if (items.length > 0) {
-        removeEvent(items[0].id)
-      }
-
-      observeDrag2(e, {
-        onMouseMove: (_e, delta) => {
-          const local = Point.add(startPos, delta)
-          const items = pianoRollStore.getNotes(local)
-          if (items.length > 0) {
-            removeEvent(items[0].id)
-          }
-        },
-      })
-    },
-  }
-}
-
-const useSelectNoteGesture = (): MouseGesture => {
-  const {
-    pianoRollStore,
-    pianoRollStore: { transform, quantizer, selectedTrack },
-    player,
-    controlStore,
-  } = useStores()
-
-  return {
-    onMouseDown(e: MouseEvent) {
-      if (selectedTrack === undefined) {
-        return
-      }
-
-      const local = pianoRollStore.getLocal(e)
-      const start = transform.getNotePoint(local)
-      const startPos = local
-
-      if (!player.isPlaying) {
-        player.position = quantizer.round(start.tick)
-      }
-
-      controlStore.selectedEventIds = []
-      pianoRollStore.selection = Selection.fromPoints(start, start)
-
-      observeDrag2(e, {
-        onMouseMove: (_e, delta) => {
-          const offsetPos = Point.add(startPos, delta)
-          const end = transform.getNotePoint(offsetPos)
-          pianoRollStore.selection = Selection.fromPoints(start, end)
-        },
-
-        onMouseUp: () => {
-          const { selection } = pianoRollStore
-          if (selection === null) {
-            return
-          }
-
-          // 選択範囲を確定して選択範囲内のノートを選択状態にする
-          // Confirm the selection and select the notes in the selection state
-          pianoRollStore.selectedNoteIds = eventsInSelection(
-            selectedTrack.events,
-            selection,
-          ).map((e) => e.id)
-
-          pianoRollStore.selection = null
-        },
-      })
-    },
-  }
-}
-
-const useRemoveNoteFromSelectionGesture = (): MouseGesture<[number]> => {
-  const {
-    pianoRollStore,
-    pianoRollStore: { selectedTrack, selectedNoteIds },
-  } = useStores()
-
-  return {
-    onMouseDown(_e: MouseEvent, noteId: number) {
-      if (selectedTrack === undefined || selectedNoteIds.length === 0) {
-        return
-      }
-
-      pianoRollStore.selectedNoteIds = selectedNoteIds.filter(
-        (id) => id !== noteId,
-      )
-    },
-  }
-}
-
-const useAddNoteToSelectionGesture = (): MouseGesture<[number]> => {
-  const { pianoRollStore } = useStores()
-
-  return {
-    onMouseDown(_e: MouseEvent, noteId: number) {
-      pianoRollStore.selectedNoteIds.push(noteId)
-    },
-  }
 }
