+++ app/src/components/PianoRoll/MouseHandler/gestures/useChangeToolGesture.ts
@@ -1,0 +1,12 @@
+import { useStores } from "../../../../hooks/useStores"
+import { MouseGesture } from "../NoteMouseHandler"
+
+export const useChangeToolGesture = (): MouseGesture => {
+  const { pianoRollStore } = useStores()
+  return {
+    onMouseDown(_e: MouseEvent) {
+      pianoRollStore.toggleTool()
+      pianoRollStore.notesCursor = "crosshair"
+    },
+  }
+}
