+++ app/src/components/PianoRoll/MouseHandler/gestures/useCreateNoteGesture.ts
@@ -1,0 +1,59 @@
+import { NoteNumber } from "../../../../entities/unit/NoteNumber"
+import { useStores } from "../../../../hooks/useStores"
+import { NoteEvent } from "../../../../track"
+import { MouseGesture } from "../NoteMouseHandler"
+import { useDragNoteCenterGesture } from "./useDragNoteEdgeGesture"
+
+export const useCreateNoteGesture = (): MouseGesture => {
+  const {
+    song: { timebase },
+    pianoRollStore,
+    pianoRollStore: { transform, selectedTrack, quantizer, newNoteVelocity },
+    pushHistory,
+  } = useStores()
+  const dragNoteCenterAction = useDragNoteCenterGesture()
+
+  return {
+    onMouseDown(e: MouseEvent) {
+      if (e.shiftKey) {
+        return
+      }
+
+      const local = pianoRollStore.getLocal(e)
+      const { tick, noteNumber } = transform.getNotePoint(local)
+
+      if (
+        selectedTrack === undefined ||
+        selectedTrack.channel == undefined ||
+        !NoteNumber.isValid(noteNumber)
+      ) {
+        return
+      }
+
+      pushHistory()
+
+      const quantizedTick = selectedTrack.isRhythmTrack
+        ? quantizer.round(tick)
+        : quantizer.floor(tick)
+
+      const duration = selectedTrack.isRhythmTrack
+        ? timebase / 8 // 32th note in the rhythm track
+        : (pianoRollStore.lastNoteDuration ?? quantizer.unit)
+
+      const note = selectedTrack.addEvent({
+        type: "channel",
+        subtype: "note",
+        noteNumber: noteNumber,
+        tick: quantizedTick,
+        velocity: newNoteVelocity,
+        duration,
+      } as NoteEvent)
+
+      if (note === undefined) {
+        return
+      }
+
+      dragNoteCenterAction.onMouseDown(e, note.id)
+    },
+  }
+}
