+++ app/src/components/PianoRoll/MouseHandler/gestures/useDragScrollGesture.ts
@@ -1,0 +1,17 @@
+import { observeDrag } from "../../../../helpers/observeDrag"
+import { useStores } from "../../../../hooks/useStores"
+import { MouseGesture } from "../NoteMouseHandler"
+
+export const useDragScrollGesture = (): MouseGesture => {
+  const { pianoRollStore } = useStores()
+  return {
+    onMouseDown(_e: MouseEvent) {
+      observeDrag({
+        onMouseMove: (e: MouseEvent) => {
+          pianoRollStore.scrollBy(e.movementX, e.movementY)
+          pianoRollStore.autoScroll = false
+        },
+      })
+    },
+  }
+}
