+++ app/src/components/PianoRoll/MouseHandler/gestures/useRemoveNoteFromSelectionGesture.ts
@@ -1,0 +1,21 @@
+import { useStores } from "../../../../hooks/useStores"
+import { MouseGesture } from "../NoteMouseHandler"
+
+export const useRemoveNoteFromSelectionGesture = (): MouseGesture<[number]> => {
+  const {
+    pianoRollStore,
+    pianoRollStore: { selectedTrack, selectedNoteIds },
+  } = useStores()
+
+  return {
+    onMouseDown(_e: MouseEvent, noteId: number) {
+      if (selectedTrack === undefined || selectedNoteIds.length === 0) {
+        return
+      }
+
+      pianoRollStore.selectedNoteIds = selectedNoteIds.filter(
+        (id) => id !== noteId,
+      )
+    },
+  }
+}
