+++ app/src/components/PianoRoll/MouseHandler/gestures/useSelectNoteGesture.ts
@@ -1,0 +1,54 @@
+import { eventsInSelection } from "../../../../actions"
+import { Point } from "../../../../entities/geometry/Point"
+import { Selection } from "../../../../entities/selection/Selection"
+import { observeDrag2 } from "../../../../helpers/observeDrag"
+import { useStores } from "../../../../hooks/useStores"
+import { MouseGesture } from "../NoteMouseHandler"
+
+export const useSelectNoteGesture = (): MouseGesture => {
+  const {
+    pianoRollStore,
+    pianoRollStore: { transform, quantizer },
+    player,
+    controlStore,
+  } = useStores()
+
+  return {
+    onMouseDown(e: MouseEvent) {
+      const local = pianoRollStore.getLocal(e)
+      const start = transform.getNotePoint(local)
+      const startPos = local
+
+      if (!player.isPlaying) {
+        player.position = quantizer.round(start.tick)
+      }
+
+      controlStore.selectedEventIds = []
+      pianoRollStore.selection = Selection.fromPoints(start, start)
+
+      observeDrag2(e, {
+        onMouseMove: (_e, delta) => {
+          const offsetPos = Point.add(startPos, delta)
+          const end = transform.getNotePoint(offsetPos)
+          pianoRollStore.selection = Selection.fromPoints(start, end)
+        },
+
+        onMouseUp: () => {
+          const { selection, selectedTrack } = pianoRollStore
+          if (selection === null || selectedTrack === undefined) {
+            return
+          }
+
+          // 選択範囲を確定して選択範囲内のノートを選択状態にする
+          // Confirm the selection and select the notes in the selection state
+          pianoRollStore.selectedNoteIds = eventsInSelection(
+            selectedTrack.events,
+            selection,
+          ).map((e) => e.id)
+
+          pianoRollStore.selection = null
+        },
+      })
+    },
+  }
+}
